ai_agent/requirements.txt
langchain[openai]==0.3.23
langchain-ollama==0.3.2
pytest==8.3.5
pydantic-settings==2.8.1
pytest-asyncio==0.26.0
redisvl==0.5.2
langchain-redis==0.2.0